Model used for sensitivity analysis
Model results were not intuitive when considered from the earlier code, so various aspects of the model code were updated. These included that promotion and public transport agents are no longer included in the final analyses, and the income calculation function for private AV ownership was fixed to reflect that the income had to be larger than the private AV cost.